refactor: 컬럼명 snake_case로 변경
